Légère modif

+++ AgenceVoyage/src/main/java/repository/ActiviteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import model.Activite;
 
-public interface ActiviteRepository extends JpaRepository<Activite, Integer>{
+public interface ActiviteRepository extends JpaRepository<Activite, Long>{
 
 }
