+++ AgenceVoyage/src/main/java/repository/PassagerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import model.Passager;
 
-public interface PassagerRepository extends JpaRepository<Passager, Integer>{
+public interface PassagerRepository extends JpaRepository<Passager, Long>{
 
 }
