+++ AgenceVoyage/src/main/java/repository/PlaneteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import model.Planete;
 
-public interface PlaneteRepository extends JpaRepository<Planete, Integer>{
+public interface PlaneteRepository extends JpaRepository<Planete, Long>{
 
 }
