+++ AgenceVoyage/src/main/java/repository/TrajetRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import model.Trajet;
 
-public interface TrajetRepository extends JpaRepository<Trajet, Integer> {
+public interface TrajetRepository extends JpaRepository<Trajet, Long> {
 
 }
